Reorganized PL support and divided into files
