+++ yucca/training/augmentation/__init__.py
@@ -1,0 +1,2 @@
+import os
+os.environ['OMP_NUM_THREADS']="1"
